[kernel][vm] Set mapping cache policy on activation
The cache policy of a VMO can change if certain conditions are met, one
of those being it having no mappings. Prior to this change the cache
policy of a VMO was extracted in a separate VMO lock acquisition prior
to installing a mapping into it. This leaves a race where a user could
change the VMO cache policy, then create another mapping to the same
region and have two mappings with different cache policies.

Change-Id: I989616b68288047d8da95c0826730b03d30dfe5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907002
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>