[fxtest][rewrite] Overhaul --fuzzy matching
It turns out tha Jaro-Winkler similarity is not a great fit for the use
case of selecting tests. The main reason for this is its emphasis on
string prefixes, which for device tests are all
"fuchsia-pkg://fuchsia.com/". This results in direct selection attempts
to return far more than intended (b/297885060).

This change overhauls fuzzy matching and implements the --fuzzy flag to
control it (matching the CLI for the old implementation).

Instead of using Jaro-Winkler, we switch to Damerau–Levenshtein
distance, a specialization of Levenshtein distance in which character
transpositions are counted as an edit distance of 1 instead of 2. This
helps with common typos in test names while more closely matching the
behavior of the old fx test.

Fixed: b/297885060
Fixed: b/295340076

Change-Id: Ic493c4407d50e6ee391b3f40df58e472b07d8b2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/908874
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>