[netstack3] Confirm reachability on TCP handshake, ACK
Provide an upper-layer confirmation to NUD that the netstack has forward
reachability when either a TCP handshake completes or at least one byte
of data is acknowledged by a remote peer. In this case, the neighbor
that we are confirming is reachable is either the peer on the TCP socket
(if link-local) or the link-local gateway to that peer.

This change causes a small regression in some of the netstack3 TCP
socket benchmarks over network-tun, about 1-4% for some of the larger
message sizes, due mainly to the cost of performing a routing table
lookup in the TCP segment receive path. This could likely be optimized
by implementing route caching in the future.

Bug: 35185
Test: netstack3-core-test
Change-Id: Ieac6f85630791b5574e4fd63e87766796ae6aa3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907603
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>