[fuchsia-async] Efficiently spawn detached tasks
For tasks that are going to be detached, using Task::spawn wraps the
future in a Fuse and a RemoteHandle that won't be used.

This change adds spawn methods to EHandle specifically for spawning
detached tasks. All of the uses of Task::spawn_on were using detach so
it has been replaced with calling EHandle::spawn_detached.

Change-Id: Ida92d411c8fe6b9af4bbeb38833f62cf85bfaa42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906306
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>