[superproject] Roll llvm-project/libc [libc] Remove hard dependencies on linux from macros headers
Currently all of this logic expects to include and link with the headers
in the `linux/` directory. This patch adds a wrapper macro to optionally
add the appropriate dependency if it exists. This is preliminary to
adding some GPU specific versions of these files.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D159021

GitOrigin-RevId: c8a6b16c3373dcb1163c115d4f218660a01d365b
Original-Revision: 43b397d92f83ad56ab44d03c882fe3ea6428faa1
Roller-URL: https://ci.chromium.org/b/8771361414143359377
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I012bc8e7a6afe9e359cbfe76d8f956649dd50e80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909653