[audio][drivers][simple-codec] Fix race when creating server instances
Previously simple-codec-server acquired the instances lock after
creating a new server instance and binding it to the dispatcher. If this
happened on a non-dispatcher thread, there was a race where the instance
could receive OnUnbound on the dispatcher before it was added to the
instances list, causing a crash (or assert failure on debug builds). To
fix this, make simple-codec-server acquire the instances lock before
creating a new instance, thus ensuring that it is always added to the
list before OnUnbound is able to run.

Fixed: 132750
Multiply: nelson-brownout-protection-test: 1000
Change-Id: Icf82ccf067f0bf2a29ec462dede4249e0cf08c47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/909561
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>