[check-licenses] Removal of flutter & python3.11
The flutter framework and prebuilts have been removed from the tree, so
none of this stuff is used anymore.

The python3.11 directory was also removed some time ago, and its absence
was causing spurious errors on local & bot runs of the check-licenses
tool.  Remove references to it as well.

Change-Id: I2b40608b728b3b92118b67dffbd9429cc7a2e025
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/908102
Reviewed-by: Jerry Belton <jcecil@google.com>
Commit-Queue: David Worsham <dworsham@google.com>