[rbe] plot download bandwidth over time
Useful for studying the remote download pressure over the
course of a build.

Usage:
reproxy_logs.sh plot_download [--interval SECONDS] log.rrpl > data.csv

Model:
Assume that over the time of a DownloadResult event, that bytes
are downloaded at a constant rate.  This is not realistic,
and is only meant to give a rough idea of download congestion.
The area-under-the-curve is equal to the total bytes downloaded.

Data:
time,bytes,concurrent

* time: regularly spaced intervals
* bytes: total number of bytes downloaded in interval (estimated)
* concurrent: number of actions trying to download outputs during the
  interval.  Some actions have multiple outputs, and those outputs
  may vary greatly in size.

Bug: 123178

Change-Id: I3d9305e3fce6085716f9d66829316c907abc8520
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/908879
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>