[ffx][overnet] Refactor ssh to allow specifying the path
This adds the ability to specify the path to the ssh command
when building ssh command lines.

This ability is used in the host_pipe creation for overnet in order
to enable specifying a different command from `ssh` when testing
so the target side of the conversation can be easily mocked.

Change-Id: I221d7bf66b4aadf95353a86bffd441ffed413f01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910953
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>