[storage][perf] Run the host benchmarks on starnix
Run the host compatible storage benchmarks against the /data and /tmp
mount points in starnix.

starnix-data is a wrapper around fxfs so comparing the benchmark results
between the 2 filesystems gives a good indication of the cost of the
starnix layer.

memfs and starnix-tmp are both in-memory filesystems. starnix-tmp is
implemented directly in starnix while memfs is in its own component.
Comparing the OpenFile and StatPath benchmarks between the 2 filesystems
gives a good indication of the FIDL overhead of memfs and other fuchsia
filesystems.

Change-Id: I6a07cc49120dac538e668e67728778a099cb09f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903004
Reviewed-by: Vickie Cheng <vickiecheng@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>