[wlan] Extends all wlandevicemonitor timeouts.
wlandevicemonitor's unit tests rely on observing the
addition or removal of a device.  In presubmits and
continuous integration, we have observed cases where this
process runs much more slowly than expected, resulting in
flakes due to timeouts.  We want to keep these timeouts
because we want some notion of which time-consuming portion
of the tests are failing if they do fail.

Bug: 132978
Test: fx test wlandevicemonitor-tests
Multiply: wlandevicemonitor-tests
Change-Id: I3c686e4b7b38b1e1208a143dcd2cd9b80c1a5cac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911897
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>