[diagnostics][fidl] Move host API to partner_internal
This API is used in `ffx inspect`, `ffx log` and other ffx plugins
that read Logs or Inspect from the device. The API is equivalent to
`fuchsia.diagnostics.ArchiveAccessor` (in the partner sdk), except
that it sends the data over a socket instead of a VMO as we can't send
VMOs to the host.

We wish to move it to partner_internal as part of having the ffx log
and ffx inspect tools in the SDK.

Change-Id: Ic7ad16b5a249d0b89c3d9cd35247adb4f1fb8ad5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/905520
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
API-Review: Christopher Johnson <crjohns@google.com>
API-Review: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>