[ffx][rcs] Add ffx side of new rcs::open_capability
This adds support in the ffx libraries for using open_capability to
request a particular capability set instead of always getting exposed.
It leaves in shims on both sides so that if you *are* asking for
exposed, it uses connect_capability to avoid disrupting existing users,
and on the other end it leaves connect_capability functions that always
pass exposed to avoid having to change a lot of existing code.

This should mean everything that currently exists should continue to
work as-is, while small uses of open can start to be used (like the new
toolbox capability).

Bug: b/292290062
Bug: b/289268551
Change-Id: I87ce21b72035cecab8ae56a7bf36bafac33eb95a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903122
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>