[fxtest][rewrite] Make error more clear
If someone adds a test to their args.gn but does not re-run the build,
the test will be added to tests.json but potentially not test-list.json.

This error message describes how to fix the problem if it happens.

Change-Id: Iaccb3c5adc41f20d93c6e552830df6f7ce3d3400
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/912020
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Cho <chok@google.com>