[test][vm] Fix race in Pager.EvictionHintsVmar
The EvictionHintsVmar test had previously been modified to account for
a race in a eviction potentially happening on a range marked DONT_NEED.
However, this handling was incompleted and only accounted for 2 out of
the 3 possible scenarios.

This change accounts for all 3 potential outcomes, and refactors the
handling into a single re-usable lambda.

Change-Id: I31ff927a051d62f20f50959a0e157054ea7fc68e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/911552
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Sage Barreda <sagebarreda@google.com>