[virtualization] Add zircon guest smoke test package
Add a new test package, zircon-guest-smoke-tests, that simply boots a
zircon guest. This is intended to be the smallest zircon guest
integration test possible.

This test has a total size of 118MiB per fx blobstats. The biggest
contributors are the zircon guest images at 35MiB for the kernel and
53MiB for the filesystem.

Change-Id: I9da89798036205ade701c1e65e72dbd2f3240cc4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/910132
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>