[fdf][fidl] Create driver FIDL with fdf transport
Creates a new "Driver" FIDL protocol in the fuchsia.driver.framework
FIDL library that will be used as a replacement for the C based
lifecycle hooks in a DFv2 driver. The mechanism to connect to this is
through a single initialize C hook to exchange the channel that will be
used to communicate this new protocol.

Bug: 130579
Change-Id: I7b33a738485cf70a7524844218a7db2f2103d927
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884201
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>