[devshell][ffx] Use build-level target.default
This makes it so that we use the default target as a build level config
for ffx as the one set in the .device file adjacent to that.
Eventually we can remove the .device file as well, but for a while it's
good to do both for when people might be switching back and forth
between older and newer revisions a bit more.

As an extra precaution, we will check that if both configurations are
present, they are the same. If they differ, we suggest to the user that
they will need to re-run `fx set-device` to resolve the ambiguity. We
will return the ffx device in the meantime though.

This would most likely happen because they had switched to an older
version, run `fx set-device`, and then switched back to a newer
version without doing it again.

It also changes the warnings for running `ffx target default` to only
happen when there's likely to be a conflict, since get should generally
do the right thing now. It now only warns on setting, and then only if
you have a device pair file created.

Bug: b/290948818
Change-Id: I17142f5324a6ec661fae1bcae4354083938dc2fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821447
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>